# backend/docs/openapi-0.4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThermoTrace API",
    "description": "API do ThermoTrace — monitoramento térmico de remessas por etiqueta NFC.\n\nVersão 0.4: acesso, perfis, cargas e recepção idempotente de evidências. Declarações de início/checkpoint/final preservadas com cadeia de integridade. Receber uma evidência não confirma o hardware nem libera uma carga. Fila Android e decodificação no servidor estão em implementação.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/perfis/modelos": {
      "get": {
        "tags": [
          "Perfis térmicos"
        ],
        "summary": "Modelos",
        "operationId": "modelos_api_v1_perfis_modelos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/perfis": {
      "get": {
        "tags": [
          "Perfis térmicos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_perfis_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "somente_ativos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Somente Ativos"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Perfis térmicos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_perfis_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovoPerfil"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/perfis/{perfil_id}": {
      "get": {
        "tags": [
          "Perfis térmicos"
        ],
        "summary": "Obter",
        "operationId": "obter_api_v1_perfis__perfil_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "perfil_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Perfil Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/perfis/{perfil_id}/aprovar": {
      "post": {
        "tags": [
          "Perfis térmicos"
        ],
        "summary": "Aprovar",
        "operationId": "aprovar_api_v1_perfis__perfil_id__aprovar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "perfil_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Perfil Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AprovarPerfil"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/perfis/{perfil_id}/retirar": {
      "post": {
        "tags": [
          "Perfis térmicos"
        ],
        "summary": "Retirar",
        "operationId": "retirar_api_v1_perfis__perfil_id__retirar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "perfil_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Perfil Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/produtos": {
      "post": {
        "tags": [
          "Produtos e configuração rápida"
        ],
        "summary": "Configurar",
        "operationId": "configurar_api_v1_produtos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 160,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurarProduto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoSaida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Produtos e configuração rápida"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_produtos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProdutoSaida"
                  },
                  "title": "Response Listar Api V1 Produtos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/produtos/localizar": {
      "get": {
        "tags": [
          "Produtos e configuração rápida"
        ],
        "summary": "Localizar",
        "operationId": "localizar_api_v1_produtos_localizar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "codigo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._/-]{0,63}$",
              "title": "Codigo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoSaida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/produtos/{produto_id}": {
      "get": {
        "tags": [
          "Produtos e configuração rápida"
        ],
        "summary": "Obter",
        "operationId": "obter_api_v1_produtos__produto_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Produto Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoSaida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AprovarPerfil": {
        "properties": {
          "declaracao_responsavel": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 20,
            "title": "Declaracao Responsavel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "declaracao_responsavel"
        ],
        "title": "AprovarPerfil"
      },
      "ConfigurarProduto": {
        "properties": {
          "codigo": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._/-]{0,63}$",
            "title": "Codigo"
          },
          "perfil_termico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Perfil Termico Id"
          },
          "intervalo_segundos": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 1.0,
            "title": "Intervalo Segundos",
            "default": 600
          },
          "versao_anterior": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Versao Anterior"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "codigo",
          "perfil_termico_id"
        ],
        "title": "ConfigurarProduto"
      },
...
      "EvidenciaPerfil": {
        "properties": {
          "produto": {
            "type": "string",
            "maxLength": 240,
            "minLength": 2,
            "title": "Produto"
          },
          "apresentacao": {
            "type": "string",
            "maxLength": 240,
            "minLength": 2,
            "title": "Apresentacao"
          },
          "condicao": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Condicao"
          },
          "fonte_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Fonte Url"
          },
          "fonte_titulo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Fonte Titulo"
          },
          "fonte_versao": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Fonte Versao"
          },
          "verificada_em": {
            "type": "string",
            "format": "date",
            "title": "Verificada Em"
          },
          "justificativa": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Justificativa"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "produto",
          "apresentacao",
          "condicao",
          "fonte_url",
          "fonte_titulo",
          "fonte_versao",
          "verificada_em",
          "justificativa"
        ],
        "title": "EvidenciaPerfil"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NovoPerfil": {
        "properties": {
          "codigo": {
            "type": "string",
            "pattern": "^[A-Z0-9][A-Z0-9_-]{1,47}$",
            "title": "Codigo"
          },
          "rotulo": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Rotulo"
          },
          "min_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 150.0,
                "minimum": -200.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,3}|(?=[\\d.]{1,6}0*$)\\d{0,3}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Min C"
          },
          "max_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 150.0,
                "minimum": -200.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,3}|(?=[\\d.]{1,6}0*$)\\d{0,3}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Max C"
          },
          "evidencia": {
            "$ref": "#/components/schemas/EvidenciaPerfil"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "codigo",
          "rotulo",
          "min_c",
          "max_c",
          "evidencia"
        ],
        "title": "NovoPerfil"
      },
...
      "ProdutoSaida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          },
          "perfil_termico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Perfil Termico Id"
          },
          "intervalo_segundos": {
            "type": "integer",
            "title": "Intervalo Segundos"
          },
          "disponivel_para_nova_carga": {
            "type": "boolean",
            "title": "Disponivel Para Nova Carga"
          },
          "produto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Produto"
          },
          "apresentacao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apresentacao"
          },
          "condicao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condicao"
          },
          "min_c": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min C"
          },
          "max_c": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max C"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "versao",
          "ativo",
          "perfil_termico_id",
          "intervalo_segundos",
          "disponivel_para_nova_carga",
          "produto",
          "apresentacao",
          "condicao",
          "min_c",
          "max_c"
        ],
        "title": "ProdutoSaida"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}